ppccap.c: Split out algorithm-specific functions
Fixes #13336
