+++ crypto/bn/build.info
@@ -77,7 +77,7 @@ IF[{- !$disabled{asm} -}]
   $BNASM_parisc20_64=$BNASM_parisc11
   $BNDEF_parisc20_64=$BNDEF_parisc11
 
-  $BNASM_ppc32=bn-ppc.s ppc-mont.s
+  $BNASM_ppc32=bn_ppc.c bn-ppc.s ppc-mont.s
   $BNDEF_ppc32=OPENSSL_BN_ASM_MONT
   $BNASM_ppc64=$BNASM_ppc32 ppc64-mont-fixed.s
   $BNDEF_ppc64=$BNDEF_ppc32
