+++ crypto/ec/ecp_nistp521.c
@@ -688,7 +688,7 @@ void p521_felem_square(largefelem out, const felem in);
 void p521_felem_mul(largefelem out, const felem in1, const felem in2);
 
 # if defined(_ARCH_PPC64)
-#  include "ppc_arch.h"
+#  include "crypto/ppc_arch.h"
 # endif
 
 void felem_select(void)
